Move packages to `internal` for better encapsulation
Refactored package structure by moving several packages into the `internal` directory to enforce encapsulation and limit their exposure to external usage. Updated all relevant import paths across the codebase to reflect this change.
