+++ cmd/api.go
@@ -10,8 +10,8 @@ import (
 	"github.com/shaharia-lab/goai"
 	"github.com/shaharia-lab/goai/mcp"
 	handlers "github.com/shaharia-lab/mcp-kit/internal/handler"
-	"github.com/shaharia-lab/mcp-kit/observability"
-	"github.com/shaharia-lab/mcp-kit/storage"
+	"github.com/shaharia-lab/mcp-kit/internal/observability"
+	"github.com/shaharia-lab/mcp-kit/internal/storage"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"go.opentelemetry.io/otel/trace"
