+++ cmd/server.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"github.com/shaharia-lab/goai/mcp"
 	goaiObs "github.com/shaharia-lab/goai/observability"
-	"github.com/shaharia-lab/mcp-kit/config"
-	"github.com/shaharia-lab/mcp-kit/observability"
-	"github.com/shaharia-lab/mcp-kit/prompt"
-	"github.com/shaharia-lab/mcp-kit/tools"
+	"github.com/shaharia-lab/mcp-kit/internal/config"
+	"github.com/shaharia-lab/mcp-kit/internal/observability"
+	"github.com/shaharia-lab/mcp-kit/internal/prompt"
+	"github.com/shaharia-lab/mcp-kit/internal/tools"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"go.opentelemetry.io/otel/codes"
