+++ cmd/task.go
@@ -7,8 +7,8 @@ import (
 	"github.com/anthropics/anthropic-sdk-go"
 	"github.com/fatih/color"
 	"github.com/shaharia-lab/goai"
-	"github.com/shaharia-lab/mcp-kit/config"
-	"github.com/shaharia-lab/mcp-kit/tools"
+	"github.com/shaharia-lab/mcp-kit/internal/config"
+	"github.com/shaharia-lab/mcp-kit/internal/tools"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
