+++ internal/prompt/registry.go
